changed toggle_privacy to set_privacy | moved usage manager out of onvif asset
